+++ packages/eyes-storybook/scripts/preconfigured.config.js
@@ -5,7 +5,7 @@ module.exports = {
   batchName: process.env.APPLITOOLS_BATCH_NAME || 'Test storybook',
   storybookUrl: process.env.APPLITOOLS_STORYBOOK_URL,
   readStoriesTimeout: 300000,
-  waitBeforeScreenshot: process.env.APPLITOOLS_WAIT_BEFORE_SCREENSHOT || 200,
+  waitBeforeCapture: process.env.APPLITOOLS_WAIT_BEFORE_CAPTURE || 200,
   disableBrowserFetching: true,
   browser: [
     //{width: 1200, height: 800, name: 'chrome'},