+++ packages/eyes-storybook/test/e2e/happy-config/visual-grid-options.config.js
@@ -6,6 +6,6 @@ module.exports = {
   storybookConfigDir: path.resolve(__dirname, '../../fixtures/customFont'),
   storybookStaticDir: path.resolve(__dirname, '../../fixtures'),
   browser: [{width: 640, height: 480, name: 'chrome'}],
-  waitBeforeScreenshot: 1000,
+  waitBeforeCapture: 1000,
   visualGridOptions: {chromeHeadless: false},
 };