+++ packages/eyes-storybook/test/fixtures/waitBeforeCapture/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+storiesOf('WaitBeforeCapture', module).add('WaitBeforeCapture page', () => {
+  return <div id="main">
+  <div id="indicator"></div>
+  <div id="text">DEFAULT</div>
+</div>
+})