+++ packages/eyes-storybook/test/it/getStories.it.test.js
@@ -104,7 +104,7 @@ describe('getStories', () => {
             parameters: {
               ...parameters,
               eyes: {
-                waitBeforeScreenshot: '.ready',
+                waitBeforeCapture: '.ready',
               },
             },
           },