+++ packages/eyes-storybook/test/unit/generateConfig.test.js
@@ -70,18 +70,25 @@ describe('generateConfig', function() {
     expect(config).to.eql({bla: 'bla from argv', ...sideEffectConfig});
   });
 
-  it('handles number waitBeforeScreenshot from env variable', () => {
-    process.env.APPLITOOLS_WAIT_BEFORE_SCREENSHOT = '1234';
+  it('handles number waitBeforeCapture from env variable', () => {
+    process.env.APPLITOOLS_WAIT_BEFORE_CAPTURE = '1234';
     const config = generateConfig({
-      externalConfigParams: ['waitBeforeScreenshot'],
+      externalConfigParams: ['waitBeforeCapture'],
     });
-    expect(config).to.eql({waitBeforeScreenshot: 1234, ...sideEffectConfig});
+    expect(config).to.eql({waitBeforeCapture: 1234, ...sideEffectConfig});
   });
 
   it('backward compatible for waitBeforeScreenshots', () => {
     process.env.APPLITOOLS_WAIT_BEFORE_SCREENSHOTS = '.someClass';
-    const defaultConfig = {waitBeforeScreenshot: 50, waitBeforeScreenshots: 50};
+    const defaultConfig = {waitBeforeCapture: 50, waitBeforeScreenshots: 50};
     const config = generateConfig({defaultConfig});
-    expect(config.waitBeforeScreenshot).to.eql('.someClass');
+    expect(config.waitBeforeCapture).to.eql('.someClass');
+  });
+
+  it('backward compatible for waitBeforeScreenshot', () => {
+    process.env.APPLITOOLS_WAIT_BEFORE_SCREENSHOT = '.someClass';
+    const defaultConfig = {waitBeforeCapture: 50, waitBeforeScreenshot: 50};
+    const config = generateConfig({defaultConfig});
+    expect(config.waitBeforeCapture).to.eql('.someClass');
   });
 });