+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- change `takeDomSnapshots` to accept function for `waitBeforeCapture`
+
 ## 12.23.21 - 2021/10/29
 
 - extend rGridResource to include and calculate a cacheKey , browserName and more to support key that is not only a resource url